+++ src/app/page.tsx
@@ -34,7 +34,7 @@ export default function Home() {
         setContainerStyles(newStyles)
       }, [view])
 
-    //todo: плохая реализация: трата ресурсов на обработку каждого скрола + утечка памяти из-за отсутствия отписок на событие   
+    //todo: плохая реализация: трата ресурсов на обработку каждого скрола   
     useEffect(() => {
         const handleScroll = () => {
             // Проверка на загрузку и достижение нижней части страницы. Добавил 10px для надежного распознавания конца страницы
