+++ src/components/ui/MyRecipies/MyRecipies.tsx
@@ -1,15 +1,10 @@
 'use client'
 
-import { FC, FormEvent, useMemo, useState } from 'react'
+import { FC, FormEvent, useState } from 'react'
 import styles from './MyRecipies.module.scss'
-import { useLazyGetUserDataQuery } from '@/store/features/user/user.actions'
 import RecipeList from '@/components/ui/RecipeList/RecipeList'
-import { TabData } from '@/components/ui/Tabs/Tabs.module'
-import Subscriptions from '@/components/ui/Subscriptions/Subscriptions'
-import Subscribers from '@/components/ui/Subscribers/Subscribers'
 import { getUseMyRecipies } from '@/hooks/useMyRecipies'
 import Button from '../Button/Button'
-import { Loader } from '../Loader/Loader'
 
 type MyRecipiesProps = {
   username?: string
@@ -17,29 +12,6 @@ type MyRecipiesProps = {
 
 const MyRecipies: FC<MyRecipiesProps> = ({ username }) => {
   const [date, setDate] = useState<string>('')
-  const [trigger, { data, error, isError, isLoading, isFetching }] =
-    useLazyGetUserDataQuery()
-
-  const tabs: TabData[] = useMemo(
-    () =>
-      username
-        ? [
-            {
-              label: `Рецепты`,
-              Content: <RecipeList dispatcher={getUseMyRecipies(username)} />,
-            },
-            {
-              label: `Мои подписки`,
-              Content: <Subscriptions />,
-            },
-            {
-              label: `Мои подписчики`,
-              Content: <Subscribers />,
-            },
-          ]
-        : [],
-    [username],
-  )
 
   const handleSelectSortByDate = () => {
     console.log('handleSelectSortByDate')
@@ -54,10 +26,6 @@ const MyRecipies: FC<MyRecipiesProps> = ({ username }) => {
     console.log('handleDateChange', event.currentTarget.value)
     setDate(event.currentTarget.value)
   }
-
-  if (error) return <div>{String(error)}</div>
-
-  if (isLoading) return <Loader />
 
   return (
     <div className={styles.container}>
@@ -77,7 +45,9 @@ const MyRecipies: FC<MyRecipiesProps> = ({ username }) => {
         />
       </div>
 
-      {username && <RecipeList dispatcher={getUseMyRecipies(username)} />}
+      {username && (
+        <RecipeList dispatcher={getUseMyRecipies(username)} view="feed" />
+      )}
     </div>
   )
 }
