+++ src/components/ui/UserCard/UserCard.module.scss
@@ -1,0 +1,34 @@
+.userContainer {
+  display: flex;
+  flex-direction: column;
+  gap: 16px;
+
+  .userFooter {
+    display: flex;
+    flex-direction: row;
+    justify-content: space-between;
+
+    button {
+      align-self: flex-start;
+      border: none;
+      background: none;
+    }
+  }
+
+  .userCard {
+    display: flex;
+    flex-direction: column;
+    gap: 24px;
+
+    h2 {
+      font-size: 18px;
+    }
+  }
+
+  .userInfo {
+    display: flex;
+    flex-direction: column;
+    gap: 12px;
+    font-size: 14px;
+  }
+}
